Remove questions table in migrations/seeds, improve migration
"Up" migrations will no longer drop existing table. Previously this
meant deploying a change would wipe the old database.

Also, drop reliance on unnecessary `is_playing` field when migrating.
This will have to be changed in the server code.

Drop table creation/seeding of the questions table as we are using
Google Sheets as the sole source of truth.

+++ server/migrations/M0001.js
@@ -1,39 +1,21 @@
-// Add tables (without foreign keys). Comments are examples of entries
 exports.up = function (knex) {
-	const schemas = [
-		{
-			// List of players with their names and company handles, and whether they're playing currently
-			table: 'players',
-			schemaFn: (t) => {
+	// .createTableIfNotExists generates a plain "CREATE TABLE IF NOT EXIST..." query 
+	// which won't work correctly if there are any alter table queries generated for 
+	// columns afterwards. Instead, use .hasTable => .createTable.
+	return knex.schema.hasTable('players').then((exists) => {
+		if (!exists) {
+			return knex.schema.createTable('players',  (t) => {
 				t.increments('id').primary();
 				t.string('handle').unique();
-				t.boolean('is_playing').defaultTo(true);
-			},
-		},
-		{
-			// List of questions with questions and hints
-			table: 'questions',
-			schemaFn: (t) => {
-				t.increments('id').primary();
-				t.text('question').notNullable();
-				t.text('hint');
-				t.timestamp('last_asked');
-			},
-		},
-	];
-
-	return Promise.all(
-		schemas.map(({ table, schemaFn }) => {
-			return knex.schema
-				.dropTableIfExists(table)
-				.then(() => knex.schema.createTable(table, schemaFn));
-		}),
-	);
+			});
+		}
+	});
 };
 
 exports.down = function (knex) {
-	return Promise.all([
-		knex.schema.dropTableIfExists('players'),
-		knex.schema.dropTableIfExists('puzzles'),
-	]);
+	if (process.env.NODE_ENV === 'production') {
+		throw new Error('Not implemented');
+	} else {
+		return knex.schema.dropTableIfExists('players');
+	}
 };
